Move Polygon modal to protocol
