+++ libs/base/src/lib/components/layout/Layout.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useLayoutEffect } from 'react'
+import React, { FC } from 'react'
 import styled, {
   createGlobalStyle,
   CSSObject,
@@ -13,9 +13,7 @@ import { ModalProvider } from 'react-modal-hook'
 
 import { Color, FontSize, Size, Spacing, ViewportWidth } from '@apps/base/theme'
 import { ReactTooltip, Tooltip } from '@apps/components/core'
-import { usePolygonModal } from '@apps/hooks'
 
-import { ChainIds, useNetwork } from '../../context/NetworkProvider'
 import { Footer } from './Footer'
 import { AppBar } from './AppBar'
 import { Toasts } from './Toasts'
@@ -233,16 +231,6 @@ const Container = styled.div`
 `
 
 export const Layout: FC = ({ children }) => {
-  const { chainId } = useNetwork()
-  const showPolygonModal = usePolygonModal()
-
-  useLayoutEffect(() => {
-    if (chainId === ChainIds.MaticMainnet && !localStorage.getItem('polygonViewed')) {
-      localStorage.setItem('polygonViewed', 'true')
-      showPolygonModal()
-    }
-  }, [chainId, showPolygonModal])
-
   return (
     <ModalProvider rootComponent={TransitionGroup}>
       <Background />
